prac02: Remove initial "number-of-lines"
Proposed problem description update (I can't update the PDF directly):

  The input contains zero or more lines of at most 80 characters, including
  a trailing newline.  Each line contains two words separated by a space.

  <Remove "3" from the Sample Input box>

Update the input data and the sample solutions to match.

Note that the C solution uses fgets() and sscanf("%80s") instead of the
original scanf("%s").  This imposes arbitrary limits on the line and word
length, but makes it safe from buffer overruns.